Update MetastoreAuditProcessor to reduce storage, also refactor some code
